+++ main/dropshippers/apps.py
@@ -1,6 +1,9 @@
 from django.apps import AppConfig
 
 
+
 class DropshippersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'dropshippers'
+
+
