+++ main/order/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class OrderConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'order'
+    
+    def ready(self):
+        import order.signals
